+++ packages/ui/core-components/src/lib/atoms/inputs/date-input/DateInput.svelte
@@ -15,6 +15,8 @@
 	import { toBoolean } from '../../../utils.js';
 	import { dateToYYYYMMDD, formatDateString } from './helpers.js';
 	import { buildQuery } from '@evidence-dev/component-utilities/buildQuery';
+	import InlineError from '../InlineError.svelte';
+	import checkRequiredProps from '../checkRequiredProps.js';
 
 	const inputs = getInputContext();
 
@@ -26,6 +28,7 @@
 	export let description = undefined;
 	/** @type {boolean} */
 	export let hideDuringPrint = true;
+	$: hideDuringPrint = toBoolean(hideDuringPrint);
 
 	/** @type {string | Date | undefined} */
 	export let start;
@@ -42,11 +45,23 @@
 	export let defaultValue;
 	/** @type {boolean| string} */
 	export let range = false;
-
 	$: range = toBoolean(range);
 
 	let query;
-	$: if (data && dates) {
+	let errors = [];
+	$: if (data) {
+		if (typeof data !== 'object') {
+			if (typeof data === 'string') {
+				errors.push(
+					`'${data}' is not a recognized query result. Data should be provided in the format: data = {'${data.replace('data.', '')}'}`
+				);
+			} else {
+				errors.push(
+					`'${data}' is not a recognized query result. Data should be an object. e.g data = {QueryName}`
+				);
+			}
+		}
+
 		const queryString = `SELECT min(${dates}) as start, max(${dates}) as end FROM ${typeof data === 'string' ? data : `(${data.text})`}`;
 		const id = `DateRange-${name}`;
 		const opts = {
@@ -57,60 +72,89 @@
 		};
 		query = buildQuery(queryString, id, opts.initialData, opts);
 		query.fetch();
+	} else if (dates) {
+		errors.push(`data is required to access a ${dates} column`);
 	}
 
 	$: startString = formatDateString(start || $query?.[0].start || new Date(0));
 	$: endString = formatDateString(end || $query?.[0].end || new Date());
 
 	$: hydrateFromUrlParam(name, (v) => {
-			if (v) {
-				if (v.includes('_to_')) {
-					// Handle date range
-					const [start, end] = v.split('_to_');
-					startString = start;
-					endString = end;
-				} else {
-					// Handle single date
-					startString = v;
-				}
+		if (v) {
+			if (v.includes('_to_')) {
+				// Handle date range
+				const [start, end] = v.split('_to_');
+				startString = start;
+				endString = end;
+			} else {
+				// Handle single date
+				startString = v;
 			}
 		}
-	);
+	});
 
 	let currentDate = dateToYYYYMMDD(new Date());
 
+	let extraDayEndString;
+
+	$: if (endString && range) {
+		extraDayEndString = new Date(endString);
+		extraDayEndString.setDate(extraDayEndString.getDate() + 1);
+		extraDayEndString = formatDateString(extraDayEndString);
+	}
+
 	function onSelectedDateInputChange(selectedDateInput) {
-	if (selectedDateInput) {
-		// Handle date range
-		if ((selectedDateInput.start || selectedDateInput.end) && range) {
-			const start = dateToYYYYMMDD(selectedDateInput.start?.toDate(getLocalTimeZone()) ?? new Date(0));
-			const end = dateToYYYYMMDD(selectedDateInput.end?.toDate(getLocalTimeZone()) ?? new Date());
-			$inputs[name] = { start, end };
+		if (selectedDateInput) {
+			// Handle date range
+			if ((selectedDateInput.start || selectedDateInput.end) && range) {
+				const start = dateToYYYYMMDD(
+					selectedDateInput.start?.toDate(getLocalTimeZone()) ?? new Date(0)
+				);
+				const end = dateToYYYYMMDD(selectedDateInput.end?.toDate(getLocalTimeZone()) ?? new Date());
+				$inputs[name] = { start, end };
 
-			// Serialize range for URL (e.g., "2024-01-01_to_2024-01-31")
-			const serializedRange = `${start}_to_${end}`;
-			updateUrlParam(name, serializedRange);
-		} 
-		// Handle single date
-		else if (!range) {
-			const value = dateToYYYYMMDD(selectedDateInput.toDate(getLocalTimeZone()) ?? new Date(0));
-			$inputs[name] = { value };
+				// Serialize range for URL (e.g., "2024-01-01_to_2024-01-31")
+				const serializedRange = `${start}_to_${end}`;
+				updateUrlParam(name, serializedRange);
+			}
+			// Handle single date
+			else if (!range) {
+				const value = dateToYYYYMMDD(selectedDateInput.toDate(getLocalTimeZone()) ?? new Date(0));
+				$inputs[name] = { value };
 
-			// Update URL with single date
-			updateUrlParam(name, value);
+				// Update URL with single date
+				updateUrlParam(name, value);
+			}
 		}
 	}
-}
 
+	let hasQueryError = false;
+	$: if ($query?.error && !hasQueryError) {
+		errors = [...errors, $query.error];
+		hasQueryError = true;
+	}
+
+	try {
+		checkRequiredProps({ name });
+	} catch (e) {
+		errors.push(e.message);
+	}
 </script>
 
 <HiddenInPrint enabled={hideDuringPrint}>
 	<div class={`${title ? '-mt-0.5' : 'mt-2'} mb-4 ml-0 mr-2 inline-block`}>
-		{#if title && range}
+		{#if title}
 			<span class="text-sm text-gray-500 block mb-1">{title}</span>
 		{/if}
 
-		{#if $query?.error}
+		{#if errors.length > 0}
+			<InlineError
+				inputType={range ? 'DateRange' : 'DateInput'}
+				height="32"
+				width={range ? 337 : 190}
+				error={errors}
+			/>
+			<!-- {:else if $query?.error}
 			<span
 				class="group inline-flex items-center relative cursor-help cursor-helpfont-sans px-1 border border-negative py-[1px] bg-negative/10 rounded"
 			>
@@ -120,7 +164,7 @@
 				>
 					{$query.error}
 				</span>
-			</span>
+			</span> -->
 		{:else}
 			<QueryLoad data={query} let:loaded>
 				<svelte:fragment slot="skeleton">
@@ -130,6 +174,7 @@
 					{onSelectedDateInputChange}
 					start={startString}
 					end={endString}
+					{extraDayEndString}
 					loaded={loaded?.ready ?? true}
 					{presetRanges}
 					{defaultValue}