+++ packages/ui/core-components/src/lib/atoms/inputs/dropdown/Dropdown.svelte
@@ -27,10 +27,12 @@
 	import HiddenInPrint from '../shared/HiddenInPrint.svelte';
 	import formatTitle from '@evidence-dev/component-utilities/formatTitle';
 	import VirtualList from './Virtual.svelte';
-	import { toBoolean } from '../../../utils.js';
 	import Info from '../../../unsorted/ui/Info.svelte';
 	import { browserDebounce } from '@evidence-dev/sdk/utils';
+	import InlineError from '../InlineError.svelte';
+	import { toBoolean } from '$lib/utils.js';
 	const inputs = getInputContext();
+	import checkRequiredProps from '../checkRequiredProps.js';
 
 	/////
 	// Component Things
@@ -47,13 +49,18 @@
 	 */
 	export let multiple = false;
 
+	$: multiple = toBoolean(multiple);
+
 	/**
 	 * When true, dropdown will not be shown during print
 	 * @type {boolean}
 	 */
 	export let hideDuringPrint = true;
+	$: hideDuringPrint = toBoolean(hideDuringPrint);
 
 	export let disableSelectAll = false;
+
+	$: disableSelectAll = toBoolean(disableSelectAll);
 
 	/**
 	 * @type {string | string[]}
@@ -61,11 +68,15 @@
 	export let defaultValue = [];
 	export let noDefault = false;
 
+	$: noDefault = toBoolean(noDefault);
+
 	/**
 	 * When true, all values will be selected by default
 	 * @type {boolean}
 	 */
 	export let selectAllByDefault = false;
+
+	$: selectAllByDefault = toBoolean(selectAllByDefault);
 
 	/**
 	 * @type {string | undefined}
@@ -213,6 +224,42 @@
 	$: open ? pauseSorting() : resumeSorting();
 
 	$: if ($finalQuery?.dataLoaded) opts = $finalQuery;
+
+	let errors = [];
+
+	if (!value) {
+		if (data) {
+			errors.push('Missing required prop: "value".');
+		} else if (!$$slots.default) {
+			errors.push('Dropdown requires either "value" and "data" props or <DropdownOption />.');
+		}
+	}
+
+	if (data) {
+		if (typeof data !== 'object') {
+			if (typeof data === 'string') {
+				errors.push(
+					`'${data}' is not a recognized query result. Data should be provided in the format: data = {'${data.replace('data.', '')}'}`
+				);
+			} else {
+				errors.push(
+					`'${data}' is not a recognized query result. Data should be an object. e.g data = {QueryName}`
+				);
+			}
+		}
+	}
+
+	try {
+		checkRequiredProps({ name });
+	} catch (err) {
+		errors.push(err.message);
+	}
+	// temp fix for multiple queryErrors be thrown
+	let hasQueryError = false;
+	$: if ($query?.error && hasQuery && !hasQueryError) {
+		errors = [...errors, $query.error];
+		hasQueryError = true;
+	}
 </script>
 
 <slot />
@@ -228,7 +275,9 @@
 
 <HiddenInPrint enabled={hideDuringPrint}>
 	<div class="mt-2 mb-4 ml-0 mr-2 inline-block">
-		{#if hasQuery && $query.error}
+		{#if errors.length > 0}
+			<InlineError inputType="Dropdown" error={errors} height="32" width="140" />
+			<!-- {:else if hasQuery && $query.error}
 			<span
 				class="group inline-flex items-center relative cursor-help cursor-helpfont-sans px-1 border border-negative py-[1px] bg-negative/10 rounded"
 			>
@@ -238,7 +287,7 @@
 				>
 					{$query.error}
 				</span>
-			</span>
+			</span> -->
 		{:else}
 			<Popover.Root bind:open>
 				<Popover.Trigger asChild let:builder>