+++ packages/ui/core-components/src/lib/atoms/inputs/dropdown/stories/Dropdown.stories.svelte
@@ -310,6 +310,15 @@
 	</Dropdown>
 </Story>
 
+<Story name="Name Error">
+	{@const data = Query.create(`SELECT id as value, tag as label from hashtags`, query)}
+	<Dropdown {data} value="value" label="label" name={undefined} />
+</Story>
+<Story name="Data as String">
+	{@const data = `Hello`}
+	<Dropdown {data} value="value" label="label" name="dataAsString" />
+</Story>
+
 <Story name="URL Parameter">
 	{@const data = Query.create(`SELECT id as value, tag as label from hashtags`, query)}
 	<Dropdown name="urlParam" {data} value="value" label="label" defaultValue={1} title="url" />