+++ packages/ui/core-components/src/lib/unsorted/viz/map/PointMap.svelte
@@ -7,6 +7,8 @@
 	import BaseMap from './_BaseMap.svelte';
 	import { Query } from '@evidence-dev/sdk/usql';
 	import { getThemeStores } from '../../../themes/themes.js';
+	import ErrorChart from '../core/ErrorChart.svelte';
+	import { toBoolean } from '$lib/utils.js';
 
 	const { resolveColorPalette } = getThemeStores();
 
@@ -67,6 +69,10 @@
 	/** @type {boolean} */
 	export let legend = true;
 
+	/** @type {boolean} */
+	export let ignoreZoom = false;
+	$: ignoreZoom = toBoolean(ignoreZoom);
+
 	/** @type {string|undefined} */
 	export let attribution = undefined;
 
@@ -76,34 +82,39 @@
 	$: isInitial = data?.hash === initialHash;
 </script>
 
-<BaseMap
-	let:data
-	{data}
-	{startingLat}
-	{startingLong}
-	{startingZoom}
-	{height}
-	{basemap}
-	{title}
-	{subtitle}
-	{legendPosition}
-	{isInitial}
-	{chartType}
-	{emptySet}
-	{emptyMessage}
-	{error}
-	{attribution}
->
-	<!-- move dispatch error outside of points to render error outside leafletmaps -->
-	<Points
+{#if !error}
+	<BaseMap
+		let:data
 		{data}
-		{lat}
-		{long}
-		colorPalette={colorPaletteStore}
-		{legendType}
+		{startingLat}
+		{startingLong}
+		{startingZoom}
+		{height}
+		{basemap}
+		{title}
+		{subtitle}
+		{legendPosition}
+		{isInitial}
 		{chartType}
-		{...$$restProps}
-		{legend}
-		on:error={(e) => (error = e.detail)}
-	/>
-</BaseMap>
+		{emptySet}
+		{emptyMessage}
+		{error}
+		{attribution}
+	>
+		<!-- move dispatch error outside of points to render error outside leafletmaps -->
+		<Points
+			{data}
+			{lat}
+			{long}
+			colorPalette={colorPaletteStore}
+			{legendType}
+			{chartType}
+			{...$$restProps}
+			{legend}
+			{ignoreZoom}
+			on:error={(e) => (error = e.detail)}
+		/>
+	</BaseMap>
+{:else}
+	<ErrorChart {error} title="Point Map" />
+{/if}