+++ packages/lib/component-utilities/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @evidence-dev/component-utilities
 
+## 4.0.8
+
+### Patch Changes
+
+- Updated dependencies [17eb07aff]
+- Updated dependencies [bcb0bf5dd]
+  - @evidence-dev/sdk@3.0.8
+
+## 4.0.7
+
+### Patch Changes
+
+- Updated dependencies [243229f46]
+  - @evidence-dev/sdk@3.0.7
+
+## 4.0.6
+
+### Patch Changes
+
+- Updated dependencies [c7547d83f]
+  - @evidence-dev/sdk@3.0.6
+
 ## 4.0.5
 
 ### Patch Changes