+++ sites/example-project/CHANGELOG.md
@@ -1,5 +1,56 @@
 # @evidence-dev/components
 
+## 3.5.48
+
+### Patch Changes
+
+- Updated dependencies [17eb07aff]
+- Updated dependencies [9179782aa]
+- Updated dependencies [bcb0bf5dd]
+- Updated dependencies [59d6755a8]
+  - @evidence-dev/sdk@3.0.8
+  - @evidence-dev/core-components@5.1.2
+  - @evidence-dev/component-utilities@4.0.8
+  - @evidence-dev/universal-sql@2.2.9
+  - @evidence-dev/tailwind@3.0.8
+
+## 3.5.47
+
+### Patch Changes
+
+- Updated dependencies [c3ff5614e]
+- Updated dependencies [243229f46]
+- Updated dependencies [0910a522c]
+- Updated dependencies [acef0cfae]
+- Updated dependencies [e91eed30c]
+- Updated dependencies [d67d7246d]
+- Updated dependencies [e93254909]
+- Updated dependencies [411fc0a16]
+- Updated dependencies [fcbe2dfb3]
+  - @evidence-dev/core-components@5.1.1
+  - @evidence-dev/sdk@3.0.7
+  - @evidence-dev/bigquery@2.0.10
+  - @evidence-dev/component-utilities@4.0.7
+  - @evidence-dev/universal-sql@2.2.8
+  - @evidence-dev/tailwind@3.0.7
+
+## 3.5.46
+
+### Patch Changes
+
+- Updated dependencies [27bca8e8e]
+- Updated dependencies [d97a4ce94]
+- Updated dependencies [f30d3ee5c]
+- Updated dependencies [c7547d83f]
+- Updated dependencies [09014326f]
+- Updated dependencies [2e506cdad]
+  - @evidence-dev/core-components@5.1.0
+  - @evidence-dev/sdk@3.0.6
+  - @evidence-dev/preprocess@6.0.3
+  - @evidence-dev/component-utilities@4.0.6
+  - @evidence-dev/universal-sql@2.2.7
+  - @evidence-dev/tailwind@3.0.6
+
 ## 3.5.45
 
 ### Patch Changes